Moved image filters used by soft inpainting into soft_inpainting.py from images.py
